requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.6.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
distlib==0.3.4
executing==0.8.3
filelock==3.7.1
identify==2.5.1
idna==3.3
ipdb==0.13.9
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pre-commit>=2.13.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
stack-data==0.3.0
toml==0.10.2
tomli==2.0.1
traitlets==5.3.0
tweepy==4.10.0
typing_extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
